requirements.txt
anyio==4.1.0
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-api-core==1.34.0
google-api-python-client==2.108.0
google-auth==1.35.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-cloud-core==1.7.3
google-cloud-storage==2.13.0
google-cloud-texttospeech==2.14.2
google-cloud-translate==2.0.1
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.5
oauthlib==3.2.2
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyTelegramBotAPI==4.14.0
python-telegram-bot==20.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
telebot==0.0.5
uritemplate==4.1.1
urllib3==2.1.0
